SoundGroup

Description

Module contains information about SoundGroups

Documentation

type SoundGroup = String

A SoundGroup is a group of sounds, to be identified by a capital letter as a standin for any one of those sounds.

parseSoundGroup :: String -> (Char, SoundGroup)

Parses a string into a Map Char SoundGroup