SoundGroup

Description

Module contains information about SoundGroups.

Read Syntax:

 G: abcd
 

where G is the symbol to represent the group, and abcd are the sounds in the group. In the future, the alternative syntax: GroupName: abcd is expected to work, where GroupName is any multicharacter identifier for the group.

Documentation

type SoundGroup = (Char, String, Maybe String)

A SoundGroup is a group of sounds, to be identified by a capital letter as a standin for any one of those sounds.

parseSoundGroup

Arguments

:: Int
The number to assign to this SoundGroup

-> String
The string to parse

-> SoundGroup
The parsed SoundGroup Tuple

Parses a string into a Map Char SoundGroup