Util

Description

Utility functions for Yasgheld.

Documentation

strip

Arguments

:: String
Input String

-> String
Output String, whitespace removed

Strips leading and trailing whitespace from the given String.

matches

Arguments

:: [SoundGroup]
The SoundGroups to check in

-> Char
The Character to check against

-> Char
The Character to check

-> Bool
True iff the characters match.

Determines if the given characters match after group resolution.